爬虫重构阶段二 (#20)
打算做的：

* （OK）把 spider 里面的缓存部分搬到后端（登录状态的缓存还保留）
* （OK）~~把 spider 的 handler 砍了~~~~（打算 handler 还保留，用于解析数据）~~把 spider 的 handler 砍了，后端的 infra::spider 砍了
* （OK）把后端的数据解析部分搬到 spider
* （OK）顺便把看不下去的代码重构了
* （OK）把后端的部分测试也砍掉或者是搬到 spider
* （OK）补点注释

Reviewed-on: http://code.qnxg.cn/qnxg/back_2026/pulls/20

+++ back_2024/src/infra/mysql/course.rs
@@ -15,8 +15,8 @@ pub struct CustomizeCourseInfo {
 
 pub async fn get_course_list(
     stu_id: &str,
-    xn: u32,
-    xq: u32,
+    xn: u16,
+    xq: u8,
 ) -> AppResult<Vec<CustomizeCourseInfo>> {
     let res = sqlx::query_as!(
         CustomizeCourseInfo,
